+++ app/resources/views/pages/dominion/op-center/show.blade.php
@@ -605,9 +605,10 @@
                                 @endphp
                                 <tr>
                                     <td>
-                                        {{ ucwords(str_replace('_', ' ', $buildingType)) }}
+                                        <span data-toggle="tooltip" data-placement="top" title="{{ $buildingHelper->getBuildingHelpString($buildingType) }}">
+                                            {{ ucwords(str_replace('_', ' ', $buildingType)) }}
+                                        </span>
                                         {!! $buildingHelper->getBuildingImplementedString($buildingType) !!}
-                                        <i class="fa fa-question-circle" data-toggle="tooltip" data-placement="top" title="{{ $buildingHelper->getBuildingHelpString($buildingType) }}"></i>
                                     </td>
                                     <td class="text-center">{{ number_format($amount) }}</td>
                                     <td class="text-center">{{ number_format((($amount / array_get($infoOp->data, "total_land", $landCalculator->getTotalLand($dominion))) * 100), 2) }}%</td>