+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes relevant to players in this project will be documented in th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 ## [Unreleased]
+### Fixed
+- Use new tooltip style for buildings from Survey Dominion
 
 ## [0.7.1-17]
 ### Fixed